# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGuard Core Engine",
    "description": "AI Agent Runtime Security Platform. Provides tool call interception, data sanitization, trust-based access control, and intent consistency detection.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/check": {
      "post": {
        "tags": [
          "check"
        ],
        "summary": "Check Tool Call",
        "operationId": "check_tool_call_api_v1_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/check/batch": {
      "post": {
        "tags": [
          "check"
        ],
        "summary": "Batch Check Tool Calls",
        "description": "Check multiple tool calls at once. Useful for planning phases.",
        "operationId": "batch_check_tool_calls_api_v1_check_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/check/explain": {
      "post": {
        "tags": [
          "check"
        ],
        "summary": "Explain Check",
        "description": "Explain what would happen for a tool call \u2014 which rules match and why.\n\nUseful for debugging and understanding why a tool call is blocked.\nDoes NOT record a trace span (dry-run mode).",
        "operationId": "explain_check_api_v1_check_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Explain Check Api V1 Check Explain Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/traces/{trace_id}": {
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "Get Trace",
        "operationId": "get_trace_api_v1_traces__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traces": {
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "List Traces",
        "operationId": "list_traces_api_v1_traces_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Traces Api V1 Traces Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchCheckRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckRequest"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "checks"
        ],
        "title": "BatchCheckRequest"
      },
      "BatchCheckResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CheckResponse"
            },
            "type": "array",
            "title": "Results"
          },
          "all_allowed": {
            "type": "boolean",
            "title": "All Allowed"
          }
        },
        "type": "object",
        "required": [
          "results",
          "all_allowed"
        ],
        "title": "BatchCheckResponse"
      },
      "CheckRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "sdk_version": {
            "type": "string",
            "title": "Sdk Version",
            "default": ""
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "default": ""
          },
          "client_trust_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Trust Level"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "tool_name"
        ],
        "title": "CheckRequest"
      },
      "CheckResponse": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "span_id": {
            "type": "string",
            "title": "Span Id",
            "default": ""
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "default": ""
          },
          "trust_level": {
            "type": "string",
            "title": "Trust Level",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "CheckResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpanResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "span_id": {
            "type": "string",
            "title": "Span Id"
          },
          "parent_span_id": {
            "type": "string",
            "title": "Parent Span Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "span_type": {
            "type": "string",
            "title": "Span Type"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "decision_reason": {
            "type": "string",
            "title": "Decision Reason"
          },
          "intent_drift_score": {
            "type": "number",
            "title": "Intent Drift Score"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "span_id",
          "parent_span_id",
          "agent_id",
          "span_type",
          "tool_name",
          "decision",
          "decision_reason",
          "intent_drift_score",
          "start_time",
          "end_time"
        ],
        "title": "SpanResponse"
      },
      "TraceResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/SpanResponse"
            },
            "type": "array",
            "title": "Spans"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "spans"
        ],
        "title": "TraceResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}